til-25-hihi/ocr3/requirements.txt
pytesseract==0.3.10
opencv-python==4.9.0.80
numpy==1.26.4
Pillow==10.3.0
fastapi==0.110.0
uvicorn==0.29.0